+++ examples/crm/src/contacts/ContactCreate.tsx
@@ -1,25 +1,12 @@
 import * as React from 'react';
-import {
-    CreateBase,
-    CreateProps,
-    TextInput,
-    ReferenceInput,
-    AutocompleteInput,
-    BooleanInput,
-    Form,
-    Toolbar,
-    required,
-    useCreateContext,
-} from 'react-admin';
-import { Card, CardContent, Divider, Box, Avatar } from '@mui/material';
+import { CreateBase, Form, Toolbar } from 'react-admin';
+import { Card, CardContent, Box, Avatar } from '@mui/material';
 
+import { ContactInputs } from './ContactInputs';
 import { Contact } from '../types';
 
-const Spacer = () => <Box width={20} component="span" />;
-
-export const ContactCreate = (props: CreateProps) => (
+export const ContactCreate = () => (
     <CreateBase
-        {...props}
         redirect="show"
         transform={(data: Contact) => ({
             ...data,
@@ -27,16 +14,9 @@ export const ContactCreate = (props: CreateProps) => (
             tags: [],
         })}
     >
-        <ContactCreateContent />
-    </CreateBase>
-);
-
-const ContactCreateContent = () => {
-    const { save, record } = useCreateContext();
-    return (
         <Box mt={2} display="flex">
             <Box flex="1">
-                <Form record={record as any} onSubmit={save}>
+                <Form>
                     <Card>
                         <CardContent>
                             <Box>
@@ -44,54 +24,7 @@ const ContactCreateContent = () => {
                                     <Box mr={2}>
                                         <Avatar />
                                     </Box>
-                                    <Box flex="1" mt={-1}>
-                                        <Box display="flex">
-                                            <TextInput
-                                                source="first_name"
-                                                validate={[required()]}
-                                            />
-                                            <Spacer />
-                                            <TextInput
-                                                source="last_name"
-                                                validate={[required()]}
-                                            />
-                                        </Box>
-                                        <Box display="flex">
-                                            <TextInput source="title" />
-                                            <Spacer />
-                                            <ReferenceInput
-                                                source="company_id"
-                                                reference="companies"
-                                            >
-                                                <AutocompleteInput optionText="name" />
-                                            </ReferenceInput>
-                                        </Box>
-                                        <Divider />
-                                        <Box mt={2} width={430}>
-                                            <TextInput
-                                                source="email"
-                                                fullWidth
-                                            />
-                                        </Box>
-                                        <Box display="flex">
-                                            <TextInput source="phone_number1" />
-                                            <Spacer />
-                                            <TextInput source="phone_number2" />
-                                        </Box>
-                                        <Divider />
-                                        <Box mt={2} width={430}>
-                                            <TextInput
-                                                source="background"
-                                                multiline
-                                                fullWidth
-                                            />
-                                            <TextInput
-                                                source="avatar"
-                                                fullWidth
-                                            />
-                                            <BooleanInput source="has_newsletter" />
-                                        </Box>
-                                    </Box>
+                                    <ContactInputs />
                                 </Box>
                             </Box>
                         </CardContent>
@@ -100,5 +33,5 @@ const ContactCreateContent = () => {
                 </Form>
             </Box>
         </Box>
-    );
-};
+    </CreateBase>
+);
